+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v1.0.1
+## 15 May 2018
+1. [](update)
+    * Some typos
+    * Better handling of twig inside Shortcode
+    * Added draggable to marker
+    * Added title for points
+    * Allowed points to be hash as well as array
+
 # v0.1.0
 ##  05/11/2018